Start taglib extension code
The taglib C bindings just support the very basic tags, that
are common to all file types. That almost fine for us. Almost. In
addition to what it offers, we will need to be able to handle at
least information about compilation-style albums.

Those are not the same in all file types. In id3v2 it cold be TPE2.
For vorbistags, ALBUMARTIST is commonly used.

Anyway. We will need some support for features in Taglib's file-
type specific features. This is a start for that.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>