Force a new commit for version magic to work
The first line of our commit message is used in the code when
git-version magic is enabled. That means that part of the message ends
up in a preprocessor macro, like this:

    #define FOO "stuff that's in the 1st line"

If we use characters that can mess that macro up (like a double quote),
that will crash our built, because code may end up like this:

    printf("foo bar %s baz\n", "garble "that" stuff");

So, extra care needs to be taken when writing those first lines.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>