Add lint make file target
This target in fact uses splint[1]. It currently only runs at -weak,
which should be enough for now. Note, that we currently get this
warning:

taggit.h: (in function strcaseeq)
taggit.h:159:13: Unrecognized identifier: strcasecmp
  Identifier used in code has not been declared. (Use -unrecog to
  inhibit warning)

I fail to see why this happens. strcasecmp() is defined in strings.h,
which we are including at the top of this file. Since I really don't
want to use -unrecog, I'll just let this slip for now.

[1] <http://www.splint.org/>

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>