Make read maps actually work
This adds the already documented behaviour that enables the user to
choose which tag type is read if more than one exist in a file via the
-R command line option.

This is currently done only for mp3 files - and boy, more or less *all*
code that deviates from "normal" TagLib-C-Binding behaviour is solely to
make meta information from mp3 files bearable.

This hasn't been tested too heavily yet, but the basics seem to work
just fine. Specific behavioural changes can be done in the future.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>