Move away from boost's program_options library
Reducing dependencies is probably a good idea with small utilities such
as this. Also, ‘program_options’ was a little weird to use. So use a BSD
‘getopt()’ now - like the C version of taggit did.

As a result, the main program looks a lot like a classic C program. But
that is to be expected with a C implementation of ‘getopt()’.

This could be re-implemented with C++ idioms at some point. But not
right now.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>