Add support for removing tags unsupported by TagLib
Actually, TagLib's PropertyMap abstraction: This layer helps removing
differences between the different file-types. This is very useful, since
it takes away exactly that job from us. Still, it doesn't support all
tags that are found in the wild.

When stripping, one might expect everything to be removed from the file,
which didn't happen before. With this change, the default behaviour is
changed to fix that. To get the old behaviour, use:

  amded -o keep-unsupported -S foo.m4a

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>