Use base64 with stings in machine readable output as well
The idea is simple: If we control with possible alphabet used with free-form
string payload, it is possible to make the machine-readable output protocol
robust against unknown string data sources.

Before, this output mode was deprecated, exactly because specially crafted
string payload could break parsers of this machine-readable protocol.

Using base64 on string data lifts this limitation.

The new ‘machine-dont-use-base64’ option can be used to revert to the
old (non-robust) behaviour.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>